[Refactor/#81] Refactor: Post 타입 및 메타데이터 인터페이스를 utils/post.ts로 이동
